Consolidate github-main.js and main.js
